✅🚚 Rename/reorganize test files
The `test/net/imap/test_imap_*` files are all client integration tests
that create a fake server socket and connect to it, testing a method or
a cluster of methods on `Net::IMAP`.  All other test files are for unit
tests for a specific class or module.
