+++ test/net/imap/test_authenticators.rb
@@ -3,7 +3,7 @@
 require "net/imap"
 require "test/unit"
 
-class IMAPAuthenticatorsTest < Test::Unit::TestCase
+class AuthenticatorsTest < Test::Unit::TestCase
 
   test "SASL::Authenticators.normalize_name" do
     authenticators = Net::IMAP::SASL::Authenticators
